+++ hooks-and-routing/src/components/smart/searchBar/SearchBar.tsx
