+++ hooks-and-routing/src/components/smart/searchBar/index.ts
